added progressbar on client. Consider adding varying lengths to both the server and the client. Planning on having a formula handle speeds as xp is accumulated.

+++ client.py
@@ -2,9 +2,11 @@ from kivy.app import App
 from kivy.uix.boxlayout import BoxLayout
 from kivy.lang import Builder
 from kivy.properties import DictProperty
+from kivy.uix.progressbar import ProgressBar
 from kivy.uix.screenmanager import ScreenManager, Screen
 import socket, pickle, json, time, threading
 from pathlib import Path
+from kivy.animation import Animation
 HOST = 'localhost'
 PORT = 1235
 
@@ -18,13 +20,21 @@ if not files.is_file():
     with open('relationships.p', 'wb') as file:
         pickle.dump({'copper ore': None,'iron ore': None,'copper ingot': 'copper ore', 'iron ingot': 'iron ore', 'copper armor': 'copper ingot', 'iron armor': 'iron ingot'}, file)
 
-def create_new_data():
+files = Path('data.p')
+if not files.is_file():
     with open('data.p', 'wb') as file:
         pickle.dump({'copper ore': 0,'iron ore': 0,'copper ingot': 0, 'iron ingot': 0, 'copper armor': 0, 'iron armor': 0}, file)
 
-files = Path('data.p')
+files = Path('xps.p')
 if not files.is_file():
-    create_new_data()
+    with open('xps.p', 'wb') as file:
+        pickle.dump({'mining': 0, 'smelting': 0, 'crafting': 0}, file)
+
+files = Path('groups.p')
+if not files.is_file():
+    with open('groups.p','wb') as file:
+        pickle.dump({'mining': ('copper ore', 'iron ore'),'smelting': ('copper ingot', 'iron ingot'),'crafting': ('copper armor', 'iron armor')}, file)
+
 client = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
 
 Builder.load_file('main.kv')
@@ -36,9 +46,14 @@ class LoginScreen(Screen):
         password = 'password'
         App.get_running_app().verify_credentials(username, password)
         self.manager.current = 'main'
-
-class MainLayout(Screen):
+class ThickProgressBar(ProgressBar):
     pass
+class MainLayout(Screen):
+    def animate(self):
+        self.ids.pb.value = 0
+        duration = App.get_running_app().get_duration()
+        anim = Animation(value=100, duration=duration)
+        anim.start(self.ids.pb)
 class WindowManager(ScreenManager):
     pass
 
@@ -48,9 +63,11 @@ class Idleize(App):
     with open('data.p', 'rb') as file:
         data = DictProperty(pickle.load(file))
     with open('relationships.p', 'rb') as file:
-        relationships = DictProperty(pickle.load(file))
+        relationships = pickle.load(file)
     with open('amounts.p', 'rb') as file:
         amounts = pickle.load(file)
+    with open('xps.p', 'rb') as file:
+        xps = DictProperty(pickle.load(file))
     player_name = 'JpJab'
     item = 'item'
     idling = False
@@ -61,6 +78,7 @@ class Idleize(App):
     def idle_thread(self):
         while True:
             while self.idling:
+                self.root.get_screen('main').animate()
                 time.sleep(1)
                 child_item = self.relationships[self.item]
                 if not child_item or self.data[child_item] - self.amounts[self.item] >= 0:
@@ -132,17 +150,18 @@ class Idleize(App):
                 self.data = new
                 print(f'dictionary saved: {App.get_running_app().data}')
         else:
-            create_new_data()
-            with open('data.p', 'rb') as file:
-                self.data = pickle.load(file)
+            with open('data.p', 'wb') as file:
+                pickle.dump({'copper ore': 0,'iron ore': 0,'copper ingot': 0, 'iron ingot': 0, 'copper armor': 0, 'iron armor': 0}, file)
+            with open('xps.p', 'wb') as file:
+                pickle.dump({'mining': 0, 'smelting': 0, 'crafting': 0})
 
-    # def on_start(self):
-    #     LoginScreen.verify(self)
-
-
+    def get_duration(self):
+        return 1
     def on_stop(self):
         with open('data.p', "wb") as file:
             pickle.dump(dict(self.data), file)
+        with open('xps.p', "wb") as file:
+            pickle.dump(dict(self.xps), file)
 
 
 if __name__ == "__main__":
